Moved stats command to team stats bundle
